Remove null-initialization of object field
When creating an object on the managed heap, The .NET runtime ensures
that all of its fields are initialized to zero before calling its
constructor.  It is therefore redundant to initialize a field manually
to null in the field declaration.

Nonetheless, such initialization is redundant anyway if the field is to
be re-assigned in the constructor.